requirements.txt
Django==4.2
psycopg2-binary==2.9.3
virtualenv==20.19.0
psycopg==3.1.9
djangorestframework==3.14.0
Pillow==9.5.0
django-rest-knox==4.2.0
django-extensions==3.2.1
django-filter==22.1